move Account out of vsr folder
